requirements.txt
torch>=2.1
numpy>=1.23
scikit-learn>=1.1
matplotlib>=3.7
PyYAML>=6.0
tqdm>=4.65
